Changed initialisation routine added two utilities in order to get the right masses
A word of warning. NATIF makes a clear distinction between "Jupiter" which
is code 599 and "Jupiter barycentre" which is code 5. This is also true for
other planets with a satellite system (i.e. 3 for Earth barycenter and 399
for Earth). Rebound's horizons.py module pulls always the single digit NAIF
code (i.e. 5 for Jupiter) unless one specifies and states the three digit
code ending in 99 (i.e. 599 for Jupiter) in the add call.
If a planet is pulled from Horizons database by it's name, rebound pulls
the mass of the planet as the total mass of the planet plus the mass of
it's satellites. Therefore for example: "model.add('Jupiter')" will add
the mass of the whole Jupiter System, while # "model.add('599')" adds
Jupiter without it's satellites.
To avoid this use utils.getNAIF("Name_Of_Planet") to pull the right code.
i.e utils.getNAIF("Jupiter") for the planet
	  utils.getNAIF("Jupiter barycenter") for the whole Jupiter system

As Horizons does not provide masses rebound defines them in the list at the
end of rebounds horizons.py. The NAIF codes ending in 99 refer to the planets,
single digits to the total mass of the planet plus its moons.