New Subroutines: 	utils.GetMass(body_or_id) 	utils.GetNAIF(body_or_id)
NATIF makes a clear distinction between "Jupiter" which
is code 599 and "Jupiter barycentre" which is code 5. This is also true for
other planets with a satellite system (i.e. 3 for Earth barycenter and 399
for Earth). Rebound's horizons.py module pulls always the single digit NAIF
code (i.e. 5 for Jupiter) unless one specifies and states the three digit
code ending in 99 (i.e. 599 for Jupiter) in the add call.
If a planet is pulled from Horizons database by it's name, rebound pulls
the mass of the planet as the total mass of the planet plus the mass of
it's satellites. Therefore for example: "model.add('Jupiter')" will add
the mass of the whole Jupiter System, while # "model.add('599')" adds
Jupiter without it's satellites.
To avoid this use utils.getNAIF("Name_Of_Planet") tAdded o pull the right code.
i.e utils.getNAIF("Jupiter") for the planet
	  utils.getNAIF("Jupiter barycenter") for the whole Jupiter system

As Horizons does not provide masses rebound defines them in the list at the
end of rebounds horizons.py. The NAIF codes ending in 99 refer to the planets,
single digits to the total mass of the planet plus its moons.

added the masses of the following bodies to horizons.py
=============================================================================
1. Metis
	 This moon has an irregular shape and measures 60×40×34 km across. A very
   rough estimate of its surface area could be placed between 5,800 and
   11,600 square kilometers (approx. 8,700). The bulk composition and mass
   of Metis are not known, but assuming that its mean density is like that
   of Amalthea (~0.86 g/cm^3), its mass can be estimated as ~3.6×10^16 kg.
2. Adrastea
	 has also an irregular shape and measures 20×16×14 km across.A surface area
   estimate would be between 840 and 1,600 (~1,200) km2. The bulk, composition
   and mass of Adrastea are not known, but assuming that its mean density is
   like that of Amalthea, around 0.86 g/cm3,its mass can be estimated at
   about 2×10^15 kg.
3. Thebe
   is irregularly shaped, with the closest ellipsoidal approximation being
	 116×98×84 km. Its surface area is probably between 31,000 and 59,000
   (~45,000) km^2. Its bulk density and mass are not known, but assuming
   that its mean density is like that of Amalthea (around 0.86 g/cm^3),its
   mass can be estimated at roughly 4.3 × 10^17 kg.