Edited mercurius to prevent test-particle encounters from numerically affecting planet positions
A common use-case for hybrid integrators is integrating large sets of massless test particles with a system of massive bodies. It is useful for the evolution of the planets starting from a single set of initial conditions to remain identical even when different sets of test particles are evolved. However, the fact that the test particles can trigger the close-encounter routine means that they do change the planet evolution even when they are massless. This update adds a flag that is determined during the close encounter prediction function that keeps track of whether the close encounters involve pairs of massive bodies or whether they are all test-particles encountering massive bodies.

The parameter "tponly_encounter" now tracks whether close-encounters are only between massless test particles and massive bodies in the system (tponly_encounter = 1) or whether there are encounters between bodies with mass (tponly_encounter=0). Also added a step before the encounter routine to save the post-kepler step state of the simulated particles (particles_post). If tponly_encounter=1, then after the test-particle encounters are resolved with ias15, the planet states are reverted back to their original end-of-step state using particles_post. This prevents the timing or frequency of test particle encounters from numerically altering the evolution of the massive bodies. If tponly_encounter=0, then the planets retain their post-step states from the close encounter routine (since they were gravitationally perturbed by the encounter!).

Specific file changes:
rebound.h: added tponly_encounter and particles_post to the definitions for mercurius
integrator_mercurius.c: implemented both in the integrator