Fixing the dual code displays
Not sure how to explain the problem but here goes. On most pages, any inline code is displayed for C and python in the same box, you just click "C" or "python" to view one or the other. On this 'Simulation Archive' page, the code for C and python were being displayed separately, and I think it was because there was a missing "=" before the "C" condition?